backend/requirements.txt
asgiref==3.4.1
cffi==1.15.1
cryptography==3.4.8
Django==3.2.20
django-filter==21.1
djangorestframework==3.14.0
importlib-metadata==4.8.3
Markdown==3.3.7
pycparser==2.21
PyMySQL==1.0.2
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.1.1
zipp==3.6.0
